Update biotek.py interfacing w/ calibration data
Previously, only a single gain value could be provided for AFU to uM conversion of data. Now, we can incorporate slope and intercept information, which is useful for more accurate calibration at lower concentrations of fluorophore. In the calibration_data csv, this information is provided as a tuple of the form (slope, intercept)